pyproject.toml
[project]
name = "maudy"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "carrascomj", email = "carrascomurielj@gmail.com" }
]
dependencies = [
    "torch>=2.3.0",
    "pyro-ppl>=1.9.1",
    "tqdm>=4.66.4",
    "typer-slim>=0.12.3",
    "maud-metabolic-models @ git+https://github.com/carrascomj/Maud#egg=3a7ee1ece4afbf445a8af5d2d550ede4eb5ec7a9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[behavior]
use-uv = true

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.7",
    "debugpy>=1.8.1",
    "ipykernel>=6.29.4",
    "plotnine>=0.13.6",
    "pytest>=8.3.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = ["maudy"]

[tool.hatch.build.targets.wheel]
packages = ["maudy"]

[project.scripts]
maudy = "maudy.cli:main"
